--- .github/workflows/main.yml
@@ -1,30 +1,0 @@
-# This workflow will run tests using node and then publish a package to GitHub Packages when a release is created
-# For more information see: https://help.github.com/actions/language-and-framework-guides/publishing-nodejs-packages
-
-name: Typescript CI
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  build-and-upload:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v2
-        with:
-          node-version: 16
-          registry-url: https://registry.npmjs.org/
-          cache: 'yarn'
-      - run: yarn install
-      - run: yarn build
-      - name: Upload built files
-        uses: actions/upload-artifact@v3.0.0
-        with:
-          # Artifact name
-          name: dist
-          # A file, directory or wildcard pattern that describes what to upload
-          path: ./dist/
