--- .github/workflows/publish.yml
@@ -1,34 +1,0 @@
-# This workflow will run tests using node and then publish a package to GitHub Packages when a release is created
-# For more information see: https://help.github.com/actions/language-and-framework-guides/publishing-nodejs-packages
-
-name: Publish
-
-on:
-  release:
-    types: [published]
-
-jobs:
-  publish-npm:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-      - uses: fregante/setup-git-user@v1
-      - uses: actions/setup-node@v2
-        with:
-          node-version: 16
-          registry-url: https://registry.npmjs.org/
-          cache: 'yarn'
-      - name: Get release
-        id: get_release
-        uses: bruceadams/get-release@v1.2.3
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      - run: yarn install
-      - run: yarn build
-      - run: yarn config set version-git-tag true
-      - run: yarn config set version-commit-hooks false
-      - run: yarn publish --new-version ${{ steps.get_release.outputs.tag_name }}
-        env:
-          NODE_AUTH_TOKEN: ${{secrets.npm_token}}
