+++ ExpertPlusStandard.dat
@@ -1,0 +1,7 @@
+{
+    "_version": "2.2.0",
+    "_notes": [],
+    "_obstacles": [],
+    "_events": [],
+    "_waypoints": []
+}
