--- package.json
@@ -1,27 +1,0 @@
-{
-  "name": "nootils",
-  "version": "0.1.0",
-  "description": "A remapper addon that brings you simple preset functions.",
-  "main": "dist/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/StormPacer/nootils"
-  },
-  "contributors": [
-    "cal117",
-    "StormPacer"
-  ],
-  "scripts": {
-    "build": "tsc"
-  },
-  "license": "MIT",
-  "typings": "dist/index.d.ts",
-  "dependencies": {
-    "swifter_remapper": "^1.0.5"
-  },
-  "devDependencies": {
-    "@tsconfig/node16": "^1.0.2",
-    "@types/node": "^17.0.16",
-    "typescript": "^4.6.2"
-  }
-}
