+++ script.ts
@@ -1,0 +1,8 @@
+import { Difficulty } from "https://deno.land/x/remapper@2.0.0/src/mod.ts" // MAKE SURE THIS IS ON THE LATEST REMAPPER VERSION!!!!!!!!!
+import {  } from "./nootils/index.ts";
+
+const map = new Difficulty("ExpertPlusLawless", "ExpertPlusStandard");
+
+// SCRIPT
+
+map.save();
