+++ scripts.json
@@ -1,0 +1,19 @@
+{
+    "allow": [
+        "all"
+    ],
+    "scripts": {
+        "rm": {
+            "cmd": "deno run --no-check script.ts",
+            "desc": "start ReMapper"
+        }
+    },
+    "watcher": {
+        "exts": [
+            "ts",
+            "js",
+            "rmmodel"
+        ],
+        "legacy": false
+    }
+}
