--- src/index.ts
@@ -1,22 +1,0 @@
-/*
-Nootils - Made by StormPacer and cal117 and curated by the community.
-Powered by Remapper
- */
-
-export * from './decor/wallBorderSize'
-export * from './decor/floatingPillars'
-export * from './decor/fire'
-export * from './decor/rain'
-export * from './decor/smoke'
-export * from './decor/snow'
-export * from './decor/water'
-
-export * from './notes/ghosty'
-export * from './notes/notetrail'
-export * from './notes/floatingdebris'
-
-export * from './utils/giveTrack'
-
-export * from './shapes/triangle'
-export * from './shapes/square'
-export * from './shapes/circle'
